chore(consumer): move package
